backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "FastAPI with OpenAI integration"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
# AI機能に必要なライブラリ
openai = "^1.0.0"       # 最新のOpenAIライブラリを使用
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"